pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gig-platform"
version = "1.0.0"
description = "A Flask-based web application connecting job posters with workers"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask==3.1.2",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-WTF==1.2.2",
    "Flask-Migrate==4.0.5",
    "python-dotenv==1.0.1",
    "geopy==2.4.1",
    "psycopg2-binary==2.9.9",
    "gunicorn==23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-flask==1.3.0",
    "black==24.10.0",
    "flake8==7.1.1",
    "mypy==1.12.0",
    "requests==2.32.3",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | venv
  | migrations
  | __pycache__
  | instance
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".venv", "venv", "migrations", "instance", "__pycache__"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--verbose --cov=src/gig_platform --cov-report=term-missing"